# index-server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Dev-connector API",
    "description": "API description format for different endpoints"
  },
  "schemes": ["http", "https"],
  "host": "localhost:80",
  "paths": {
...
    "/api/profiles": {
      "post": {
        "summary": "Create or Edit profile",
        "tags": ["Profile"],
        "description": "Create or edit profile",
        "parameters": [{
          "$ref": "#/parameters/Profile"
        }],
        "responses": {
          "200": {
            "description": "The profile has been created or updated",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "The request isn't according to the form principles",
            "schema": {
              "type": "object",
              "properties": {
                "handle": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "skills": {
                  "type": "string"
                },
                "website": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/current": {
      "get": {
        "summary": "Get Profile",
        "tags": ["Profile"],
        "description": "Get Current Users Profile",
        "responses": {
          "200": {
            "description": "The profile has been returned",
            "schema": {
              "type": "object",
              "allOf": [{
                "$ref": "#/definitions/Profile"
              },{
                "description": "User field populated",
                "type": "object",
                "properties": {
                  "user": {
                    "$ref": "#/definitions/UserPopulate"
                  }
                }
              }]
            }
          },
          "404": {
            "description": "User Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "noprofile": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/all": {
      "get": {
        "summary": "Get all profiles",
        "tags": ["Profile"],
        "description": "Get all profiles by handle",
        "responses": {
          "200": {
            "description": "All profiles are returned",
            "schema": {
              "type": "array",
              "items": {
                "allOf": [{
                  "$ref": "#/definitions/Profile"
                },{
                  "description": "User field populated",
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/definitions/UserPopulate"
                    }
                  }
                }]
              }
            }
          },
          "404": {
            "description": "No profiles were returned",
            "schema": {
              "type": "object",
              "properties": {
                "noprofile": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/handle/{handle}": {
      "get": {
        "summary": "Get profile",
        "tags": ["Profile"],
        "description": "Get profile by handle",
        "parameters": [{
          "name": "handle",
          "in": "path",
          "required": true,
          "description": "Handle of the profile",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "The user is being returned",
            "schema": {
              "allOf": [{
                "$ref": "#/definitions/Profile"
              },{
                "type": "object",
                "properties": {
                  "user": {
                    "$ref": "#/definitions/UserPopulate"
                  }
                }
              }]
            }
          },
          "404": {
            "description": "There is no profile for this handle",
            "schema": {
              "type": "object",
              "properties": {
                "noprofile": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/user/{userID}": {
      "get": {
        "summary": "Get Profile",
        "tags": ["Profile"],
        "description": "Get profile by user ID",
        "parameters": [{
          "name": "userID",
          "in": "path",
          "type": "string",
          "required": true,
          "description": "user ID"
        }],
        "responses": {
          "200": {
            "description": "The Profile has been returned",
            "schema": {
              "allOf": [{
                "$ref": "#/definitions/Profile"
              },{
                "type": "object",
                "properties": {
                  "user": {
                    "$ref": "#/definitions/UserPopulate"
                  }
                }
              }]
            }
          }
        }
      }
    },
    "/api/profiles/experience": {
      "post": {
        "summary": "Add experience",
        "tags": ["Profile"],
        "description": "Add experience to existing profile",
        "parameters": [{
          "$ref": "#/parameters/ProfileExperience"
        }],
        "responses": {
          "200": {
            "description": "The profile experience has been added",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid experience form params",
            "schema": {
              "properties": {
                "title": {
                  "type": "string"
                },
                "company": {
                  "type": "string"
                },
                "from": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/education": {
      "post": {
        "summary": "Add Education",
        "tags": ["Profile"],
        "description": "Add education to user profile",
        "parameters": [{
          "$ref": "#/parameters/ProfileEducation"
        }],
        "responses": {
          "200": {
            "description": "Education has been added to profile",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid education form params",
            "schema": {
              "properties": {
                "school": {
                  "type": "string"
                },
                "degree": {
                  "type": "string"
                },
                "fieldofstudy": {
                  "type": "string"
                },
                "from": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "description": "JSON WEB TOKEN",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [{
    "JWT": []
  }],
  "definitions": {
...
    "Profile": {
      "type": "object",
      "required": ["_id","user", "handle", "status", "skills"],
      "properties": {
        "_id": {
          "type": "string",
          "format": "object"
        },
        "user": {
          "type": "string",
          "format": "object"
        },
        "handle": {
          "type": "string",
          "maxLength": 40
        },
        "company": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "bio": {
          "type": "string"
        },
        "githubusername": {
          "type": "string"
        },
        "experience": {
          "$ref": "#/definitions/ProfileExperience"
        },
        "education": {
          "$ref": "#/definitions/ProfileEducation"
        },
        "social": {
          "type": "object",
          "properties": {
            "youtube": {
              "type": "string"
            },
            "twitter": {
              "type": "string"
            },
            "facebook": {
              "type": "string"
            },
            "linkedin": {
              "type": "string"
            },
            "instagram": {
              "type": "string"
            }
          }
        },
        "date": {
          "type": "string",
          "format": "date"
        },
        "__v": {
          "type": "string"
        }
      }
    },
    "ProfileExperience": {
      "type": "array",
      "items": {
        "required": ["title", "company", "from", "current"],
        "properties": {
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "current": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        }
      }
    },
    "ProfileEducation": {
      "type": "array",
      "items": {
        "required": ["school", "degree", "fieldofstudy", "from", "current"],
        "properties": {
          "school": {
            "type": "string"
          },
          "degree": {
            "type": "string"
          },
          "fieldofstudy": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "current": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        }
      }
    },
    "UserPopulate": {
      "type": "object",
      "required": ["_id"],
      "properties": {
        "_id": {
          "type": "string",
          "format": "object"
        },
        "name": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    }
  },
  "parameters": {
...
    "Profile": {
      "description": "Create Profile Data",
      "name": "ProfileBody",
      "in": "body",
      "required": true,
      "schema": {
        "type": "object",
        "required": ["handle", "status", "skills"],
        "properties": {
          "handle": {
            "type": "string",
            "maxLength": 40
          },
          "company": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "githubusername": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedin": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      }
    },
    "ProfileExperience": {
      "description": "Experience params",
      "name": "experience",
      "in": "body",
      "schema": {
        "type": "object",
        "required": ["title", "company", "from"],
        "properties": {
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "current": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        }
      }
    },
    "ProfileEducation": {
      "description": "Education params",
      "name": "education",
      "in": "body",
      "schema": {
        "type": "object",
        "required": ["school", "degree", "fieldofstudy", "from"],
        "properties": {
          "school": {
            "type": "string"
          },
          "degree": {
            "type": "string"
          },
          "fieldofstudy": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "current": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}